backend/requirements.txt
flask
flask-cors
flask-limiter
openai
python-dotenv
-e ../prepper-cli
